feat: refactor index.js into small components
