+++ src/pages/Components/Features/Features.constants.ts
@@ -1,0 +1,22 @@
+const text = {
+  features: [
+      {
+        title: "Infrastructure",
+        link: "infrastructure/azure/core_infrastructure",
+        description: "Find out how to configure your cloud provider and other core infrastructure to make Stacks work for you."
+      },
+      {
+        title: "Test Framework",
+        link: "testing/cli_testing_frameworks",
+        description: "With just one command, discover how to tailor your project with Stacks’ front-end focused opinionated test frameworks."
+      },
+  ],
+  contribute: {
+    title: "Contribute to Stacks",
+    link: "contributions",
+    description: "As an open source project, you can be a part of the Github community shaping Amido Stacks"
+  }
+}
+
+
+export default text;
