+++ src/pages/Components/Features/Features.module.css
@@ -1,0 +1,110 @@
+
+.feature h3 {
+  font-size: max(28px, min(3vw, 44px));
+  text-align: center;
+}
+
+.features {
+
+  padding: 3rem 0 2rem;
+}
+
+.features .feature {
+  margin-bottom: 3rem;
+}
+
+.features .feature p {
+  display: flex;
+  flex: 1;
+  text-align: center;
+  align-self: stretch;
+  padding: 0px 30px 0px 30px;
+}
+
+.features > .feature .buttons {
+  justify-self: end;
+  height: 46px;
+}
+
+@media (min-width: 768px) { 
+  .features {
+    padding: 0px;
+    display: flex;
+    align-items: center;
+    width: 100%;
+    flex-direction: row;
+    gap: 30px;
+    flex-wrap: wrap;
+  }
+
+  .features .feature {
+    display: flex;
+    width: calc(50% - 15px);
+    flex-grow: 1;
+    padding: 30px 0px 0px;
+    font-size: max(18px, min(2vw, 30px));
+  }
+
+
+  .features .feature div {
+    max-width: 480px;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-self: stretch;
+  }
+}
+
+.features2 {
+  flex-wrap: wrap;
+  display: flex;
+  padding: 1rem 0;
+  width: 100%;
+  flex-direction: row;
+  gap: 30px;
+}
+
+
+.features2 .feature {
+  display: flex;
+  flex-direction: column;
+  width: calc(33% - 30px);
+  flex-grow: 1;
+  padding: 30px 15px 30px 15px;
+  min-width: 320px;
+  max-width: 590px;
+  font-size: 18px;
+}
+
+.features2 .feature p {
+  display: flex;
+  flex-grow: 1;
+  font-size: max(18px, min(2vw, 30px));
+  align-self: stretch;
+}
+
+.features2 .feature h3 {
+  display: flex;
+  color: var(--ifm-heading-color);
+}
+
+.features2 .buttons a:hover {
+  color: white;
+}
+
+.mainHolder {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  background-color: #f4f4f4;
+}
+
+.main {
+  width: 100%;
+  max-width: 1600px;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+}
