+++ src/pages/Components/Introduction/Introduction.constants.ts
@@ -1,0 +1,10 @@
+const introduction = {
+    benefitsList: [
+        'Increase project predictability with reliable, tried and tested code.',
+        'Slash configuration time and speeds up deployment.',
+        'An ever-growing calalogue of workload templates, package management and example implementations.',
+        'A solid foundation for you to springboard into build.'
+    ],
+}
+
+export default introduction;
