+++ src/pages/Components/Introduction/Introduction.module.css
@@ -1,0 +1,18 @@
+.learnAboutStacks {
+    margin-top: 30px;
+    margin-bottom: 30px;
+}
+
+.intro {
+    width: 100%;
+}
+
+.intro ul li {
+    padding: 0.5rem 0;
+}
+
+@media (min-width: 768px) {
+    flex-grow: 0;
+    flex-shrink: 0;
+    flex-basis: 55%;
+}
