+++ src/pages/Components/ResponsivePlayer/ResponsivePlayer.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactPlayer from "react-player";
+
 import styles from "./styles.module.css";
 
 function ResponsivePlayer({url}) {
