+++ src/pages/Components/TechSupportedByStacks/Logos/AmidoStacksLogo.tsx
