+++ src/pages/Components/TechSupportedByStacks/Logos/AppliToolsLogo.tsx
