+++ src/pages/Components/TechSupportedByStacks/Logos/AzureFunctionsLogo.tsx
