+++ src/pages/Components/TechSupportedByStacks/Logos/CosmosDbLogo.tsx
