+++ src/pages/Components/TechSupportedByStacks/Logos/CucumberLogo.tsx
