+++ src/pages/Components/TechSupportedByStacks/Logos/CypressLogo.tsx
