+++ src/pages/Components/TechSupportedByStacks/Logos/GatlingLogo.tsx
