+++ src/pages/Components/TechSupportedByStacks/Logos/JavaLogo.tsx
