+++ src/pages/Components/TechSupportedByStacks/Logos/KarateLogo.tsx
