+++ src/pages/Components/TechSupportedByStacks/Logos/NodeLogo.tsx
