+++ src/pages/Components/TechSupportedByStacks/Logos/ReactLogo.tsx
