+++ src/pages/Components/TechSupportedByStacks/Logos/SerenityLogo.tsx
