+++ src/pages/Components/TechSupportedByStacks/Logos/SwaggerLogo.tsx
