+++ src/pages/Components/TechSupportedByStacks/Logos/TerraformLogo.tsx
