+++ src/pages/Components/TechSupportedByStacks/Logos/TestcafeLogo.tsx
