+++ src/pages/Components/WhatStacksCanDo/WhatStacksCanDo.constants.ts
@@ -1,0 +1,6 @@
+const whatStacksCanDo = {
+    clientFeedbackMessage: 'Ensono Stacks made our transformation significantly easier and gave us confidence right from day one" - '
+    
+}
+
+export default whatStacksCanDo;
