--- src/pages/Constants/Contribute.constants.ts
@@ -1,7 +1,0 @@
-const contribute = {
-    title: "Contribute to Stacks",
-    link: "contributions",
-    description: "As an open source project, you can be a part of the Github community shaping Amido Stacks"
-}
-
-export default contribute;
