--- src/pages/Constants/StacksFeatures.constants.ts
@@ -1,16 +1,0 @@
-const features = 
-    [
-        {
-          title: "Infrastructure",
-          link: "infrastructure/azure/core_infrastructure",
-          description: "Find out how to configure your cloud provider and other core infrastructure to make Stacks work for you."
-        },
-        {
-          title: "Test Framework",
-          link: "testing/cli_testing_frameworks",
-          description: "With just one command, discover how to tailor your project with Stacks’ front-end focused opinionated test frameworks."
-        },
-      ];
-
-
-export default features;
